Move `UUID` digestible implementation to `UUID` companion object
